added backend code

+++ src/main/java/com/example/DBPROJECT/entity/ParkingAreaTransactionHistory.java
@@ -21,7 +21,7 @@ public class ParkingAreaTransactionHistory {
 
     private int ContractedMerchantID;
 
-    private BigDecimal Extension;
+    private int Extension;
 
     private Timestamp EntryDate;
